+++ db/convenience.cc
@@ -58,11 +58,13 @@ Status VerifySstFileChecksum(const Options& options,
   std::unique_ptr<RandomAccessFileReader> file_reader(
       new RandomAccessFileReader(std::move(file), file_path));
   const bool kImmortal = true;
+  ColumnFamilyOptions cf_options(options);
+  MutableCFOptions mutable_cf_options(cf_options);
   s = ioptions.table_factory->NewTableReader(
-      TableReaderOptions(ioptions, options.prefix_extractor.get(), env_options,
-                         internal_comparator, false /* skip_filters */,
-                         !kImmortal, false /* force_direct_prefetch */,
-                         -1 /* level */),
+      TableReaderOptions(
+          ioptions, mutable_cf_options, options.prefix_extractor.get(),
+          env_options, internal_comparator, false /* skip_filters */,
+          !kImmortal, false /* force_direct_prefetch */, -1 /* level */),
       std::move(file_reader), file_size, &table_reader,
       false /* prefetch_index_and_filter_in_cache */);
   if (!s.ok()) {