+++ table/table_reader_bench.cc
@@ -143,7 +143,7 @@ void TableReaderBenchmark(Options& opts, EnvOptions& env_options,
     std::unique_ptr<RandomAccessFileReader> file_reader(
         new RandomAccessFileReader(std::move(raf), file_name));
     s = opts.table_factory->NewTableReader(
-        TableReaderOptions(ioptions, moptions.prefix_extractor.get(),
+        TableReaderOptions(ioptions, moptions, moptions.prefix_extractor.get(),
                            env_options, ikc),
         std::move(file_reader), file_size, &table_reader);
     if (!s.ok()) {